requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
python-dotenv
huggingface_hub